pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "picture-analyzer"
version = "2.0.0"
description = "AI-powered photo analysis, enhancement, and metadata embedding tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Leen" },
]
keywords = ["photo", "image", "analysis", "ai", "openai", "exif", "metadata", "enhancement"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "openai>=2.0",
    "Pillow>=10.0",
    "piexif>=1.1",
    "requests>=2.28",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.10"]
web = ["flask>=3.0"]
yaml = ["PyYAML>=6.0"]
numpy = ["numpy>=1.24"]
rich = ["rich>=13.0"]
all = [
    "pillow-heif>=0.10",
    "flask>=3.0",
    "PyYAML>=6.0",
    "numpy>=1.24",
    "rich>=13.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1",
]

[project.scripts]
picture-analyzer = "picture_analyzer.cli.app:main"
description-editor = "picture_analyzer.web.app:main"

[project.urls]
Repository = "https://github.com/leen/picture-analyzer"

# --- Tool Configuration ---

[tool.hatch.build.targets.wheel]
packages = ["src/picture_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"

[tool.ruff]
src = ["src"]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["picture_analyzer"]
